♻️ REFACTOR: 백통신 API 권한별 파일 분리
