+++ src/api/UserService.js
@@ -2,111 +2,11 @@ import axios from "axios";
 
 axios.defaults.baseURL = "http://localhost:8080";
 
-export const CreateContract = async (requestContractDTO) => {
-  try {
-    const token = localStorage.getItem("accessToken");
-
-    const response = await axios.post(
-      "/contract-manager/contract",
-      requestContractDTO,
-      {
-        headers: {
-          Authorization: `Bearer ${token}`,
-        },
-      }
-    );
-
-    if (response.data.success) {
-      alert("계약생성 성공");
-
-      return response.data.data;
-    } else {
-      return false;
-    }
-  } catch (error) {
-    console.error("ERROR: ", error.response.data);
-    alert(error.response.data.message);
-
-    return false;
-  }
-};
-
-export const CreateServiceDetail = async (requestContractDTO) => {
-  try {
-    const token = localStorage.getItem("accessToken");
-
-    const response = await axios.post("/evaluation-item", requestContractDTO, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    if (response.data.success) {
-      return response.data.success;
-    } else {
-      return false;
-    }
-  } catch (error) {
-    console.error("ERROR: ", error.response.data);
-
-    return false;
-  }
-};
-
-export const fetchContractInfo = async (contractId) => {
-  try {
-    const token = localStorage.getItem("accessToken");
-
-    const response = await axios.get(`/common/contract/${contractId}`, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    console.log(response);
-
-    if (response.data.success) {
-      console.log(response);
-
-      return response.data.data;
-    } else {
-      return [];
-    }
-  } catch (error) {
-    console.error(
-      "ERROR: ",
-      error.response ? error.response.data : error.message
-    );
-  }
-};
-
 export const fetchServiceInfo = async (evaluationItemId) => {
   try {
     const token = localStorage.getItem("accessToken");
 
     const response = await axios.get(`/detail/${evaluationItemId}`, {
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-
-    if (response.data.success) {
-      console.log(response.data.data);
-
-      return response.data.data;
-    } else {
-      return [];
-    }
-  } catch (error) {
-    console.error("ERROR: ", error.response.data);
-  }
-};
-
-export const fetchAllContractInfo = async () => {
-  try {
-    const token = localStorage.getItem("accessToken"); // 토큰 가져오기
-
-    const response = await axios.get("/contract-manager/all-contract", {
       headers: {
         Authorization: `Bearer ${token}`,
       },
